+++ app/Http/Controllers/ProjectController.php
@@ -4,23 +4,23 @@ namespace App\Http\Controllers;
 
 use App\Models\Project;
 use App\Models\Tahap;
-use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
     public function index()
     {
-        $projects = Project::with( 'tahap')->get();
+        $projects = Project::with('tahap')->where('id_user', Auth::id())->get();
         $tahaps = Tahap::all();
 
-        return view('projects.index', compact('projects',  'tahaps'));
+        return view('projects.index', compact('projects', 'tahaps'));
     }
 
     public function create()
     {
         $tahaps = Tahap::all();
-        return view('projects.create', compact( 'tahaps'));
+        return view('projects.create', compact('tahaps'));
     }
 
     public function store(Request $request)
@@ -32,25 +32,35 @@ class ProjectController extends Controller
             'tahap_id' => 'required|exists:tahaps,id',
         ]);
 
-        Project::create($request->all());
+        $project = new Project($request->all());
+        $project->id_user = Auth::id();
+        $project->save();
 
         return redirect()->route('projects.index')->with('success', 'Project berhasil ditambahkan');
     }
 
     public function show(Project $project)
     {
+        if ($project->id_user != Auth::id()) {
+            abort(403);
+        }
         return view('projects.show', compact('project'));
     }
 
     public function edit(Project $project)
     {
-       
+        if ($project->id_user != Auth::id()) {
+            abort(403);
+        }
         $tahaps = Tahap::all();
         return view('projects.edit', compact('project', 'tahaps'));
     }
 
     public function update(Request $request, Project $project)
     {
+        if ($project->id_user != Auth::id()) {
+            abort(403);
+        }
         $request->validate([
             'nama' => 'required',
             'nama_project' => 'required|string|max:255',
@@ -65,6 +75,9 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
+        if ($project->id_user != Auth::id()) {
+            abort(403);
+        }
         $project->delete();
         return redirect()->route('projects.index')->with('success', 'Project berhasil dihapus');
     }
